+++ src/briefy/common/db/model.py
@@ -48,6 +48,7 @@ class Security:
 class Base(Security):
     """Base Declarative model."""
 
+    __parent_attr__ = None
     __session__ = None
     __exclude_attributes__ = [
         '_sa_instance_state', 'request', 'versions', 'path', 'can_view', 'type', 'local_roles'